Move test code into separate file
