+++ tests/test_ynab.py
@@ -3,7 +3,6 @@ import unittest
 from collections import defaultdict
 
 from mock import patch, MagicMock
-import yaml
 
 from ynab import ynab
 from ynab import keyring_secrets
@@ -45,17 +44,6 @@ def concatd(*args):
     for d in args:
         ret.update(d)
     return ret
-
-class TestConfig(unittest.TestCase):
-    ''' Tests the parsing of configuration files '''
-
-    def test_parse_example_config(self):
-        script_dir = os.path.dirname(os.path.realpath(__file__))
-        example_config_file = os.path.join(script_dir, 'ynab.conf')
-        with open(example_config_file) as f:
-            as_yaml = yaml.load(f)
-        validated_yaml = ynab.parse_config(as_yaml)
-        self.assertEqual(as_yaml, validated_yaml)
 
 
 class TestGetSecretsFromKeyring(unittest.TestCase):
